press space to spin
