+++ turnaround/src/components/Button.tsx
@@ -5,8 +5,10 @@ type Props = {
     text: string
 }
 
-export const Button = ({text}: Props) => {
+const Button = ({text}: Props) => {
     return (
         <button className={styles.btn}>{text}</button>
     )
 }
+
+export default Button
