+++ turnaround/src/components/SlotLeaderboard.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 
-export const SlotLeaderboard = () => {
+const SlotLeaderboard = () => {
     return (
         <div>
             This is the leaderboard of the top winners of the slot
         </div>
     );
 }
+
+export default SlotLeaderboard
