+++ turnaround/src/components/SlotLog.tsx
@@ -1,7 +1,22 @@
 import React from "react";
 
-export const SlotLog = () => {
+type Props = {
+    wins: Array<number>
+}
+
+const SlotLog = (props: Props) => {
     return(
-        <div>This is the slot log of recent spins</div>
+        <div>
+            This is the slot log of recent spins
+            <div className={"winList"}>
+                {props.wins.map((w, i) => (
+                    <div className={"win"} key={i}>
+                        You won: {w}
+                    </div>
+                ))}
+            </div>
+        </div>
     )
 }
+
+export default SlotLog
