+++ turnaround/src/components/styles/slotControls.module.css
@@ -1,0 +1,5 @@
+.slotControls{
+    display: flex;
+    flex-direction: row;
+    gap: 1em;
+}
