+++ turnaround/src/pages/Slot.tsx
@@ -1,0 +1,13 @@
+import Slot from "../components/Slot";
+import {Header} from "../components/Header";
+
+export const SlotPage = () => {
+    return (
+        <>
+            <Header/>
+            <Slot/>
+        </>
+    )
+}
+
+export default SlotPage
